backend/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
